--- a/crates/jiff-icu/src/error.rs
+++ b/crates/jiff-icu/src/error.rs
@@ -43,9 +43,8 @@ impl core::fmt::Display for Error {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::error::Error for Error {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+impl core::error::Error for Error {
+    fn source(&self) -> Option<&(dyn core::error::Error + 'static)> {
         match self.kind {
             ErrorKind::Adhoc(_) => None,
             ErrorKind::Jiff(ref err) => Some(err),