The ICU4X project seems to be slowly working toward a 2.0 release. It
is currently on `beta.2`. There are some nice improvements here, in
particular, the ability to localize zoned datetimes.

This PR migrates to ICU4X 2.0 (first commit) and then adds support for
conversions between time zones, offsets and zoned datetimes (second
commit). I've also added a basic example demonstrating how to format a
`jiff::Zoned` value.

I've also added support for `%c`, `%r`, `%X` and `%x` to `jiff::fmt::strtime`.
Since I have no plans for Jiff to do localization, and since these specifiers
are commonly implemented via localization, I've added some
infrastructure to `jiff::fmt::strtime` that enables overriding the behavior
of these specifiers. Specifically, in Biff (not yet released), I've used this
trait with ICU4X to provide localized datetime formatting.

I'm somewhat conflicted about adding a hook for localization into
`strtime`-style APIs. In particular, the ideal way to do localization is
just by using ICU4X directly. By doing it through `strftime`, you're
somewhat limiting yourself quite a bit. But it can be quite
convenient, and it can be important if one is trying to build or
integrate with legacy APIs like POSIX. To mitigate this, I've added
documentation warning folks against using this functionality, and
instead pointing them toward ICU4X.

I also did some work optimizing `strftime` in this PR, spurred by
https://github.com/uutils/coreutils/issues/7852.

Here's the improvement from this PR for Jiff:

```
$ critcmp base change1 -f jiff
group                          base                                   change1
-----                          ----                                   -------
parse/strptime/oneshot/jiff    1.01     59.9±0.88ns        ? ?/sec    1.00     59.2±0.71ns        ? ?/sec
print/strftime/oneshot/jiff    1.74    174.7±1.45ns        ? ?/sec    1.00    100.6±0.39ns        ? ?/sec
```

And relative to `chrono` and `time` (a little tricky to read this output since we distinguish between `prebuilt` and `oneshot`, and it is appropriate to compare Jiff's `oneshot` with Chrono's `prebuilt`, since Jiff doesn't have a `prebuilt` API):

```
$ critcmp change1 -g '(.*)/(?:humantime|jiff|chrono|time)$'
group                      change1//chrono                        change1//jiff                          change1//time
-----                      ---------------                        -------------                          -------------
parse/strptime/oneshot     2.86    169.2±2.63ns        ? ?/sec    1.00     59.2±0.71ns        ? ?/sec
parse/strptime/prebuilt    1.00     81.6±0.93ns        ? ?/sec                                           1.36    110.9±0.85ns        ? ?/sec
print/strftime/oneshot     2.13    214.7±3.63ns        ? ?/sec    1.00    100.6±0.39ns        ? ?/sec
print/strftime/prebuilt    1.00     89.3±0.77ns        ? ?/sec                                           1.14    101.7±0.87ns        ? ?/sec
```